+++ config/lando.js
@@ -36,7 +36,15 @@ module.exports = {
   // Contributors pages
   contributorsPage: {
     auto: true,
+    exclude: [
+      'dependabot[bot]',
+    ],
   },
+
+  // Exclude from git contributors
+  contributorsExclude: [
+    'dependabot[bot]',
+  ],
 
   // Use Google Analytics
   ga: {