+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 v1.0.0-beta.28 [March 15, 2021](https://github.com/lando/vuepress-theme-default-plus/releases/tag/v1.0.0-beta.28)
 -------------------------------
 
-* Added support for "Reading Mode"
+* Added toggleable "Reading Mode" feature
+* Added `BlogPost` as a new page type
+* Added ability to easily add more "page types"
+* Added new "highlight" custom container
 * Improved global and page-by-page component disabling so its more consistent
+* Improved consistency of `border-radius`
 * Fixed bug in `sitemp.xml` generation
-
 
 v1.0.0-beta.27 [March 13, 2021](https://github.com/lando/vuepress-theme-default-plus/releases/tag/v1.0.0-beta.27)
 -------------------------------